chore(refactor): move util function to util package
